backend/requirement.txt
flask
pymongo
dnspython
python-dotenv


